Move some files into configs directory
