+++ core/os/os.cpp
@@ -30,12 +30,12 @@
 
 #include "os.h"
 
+#include "configs/version_generated.gen.h"
 #include "core/os/dir_access.h"
 #include "core/os/file_access.h"
 #include "core/os/input.h"
 #include "core/os/midi_driver.h"
 #include "core/project_settings.h"
-#include "core/version_generated.gen.h"
 #include "servers/audio_server.h"
 
 #include <stdarg.h>
