+++ editor/editor_about.cpp
@@ -31,11 +31,11 @@
 #include "editor_about.h"
 #include "editor_node.h"
 
+#include "configs/version_hash.gen.h"
 #include "core/authors.gen.h"
 #include "core/donors.gen.h"
 #include "core/license.gen.h"
 #include "core/version.h"
-#include "core/version_hash.gen.h"
 
 // The metadata key used to store and retrieve the version text to copy to the clipboard.
 static const String META_TEXT_TO_COPY = "text_to_copy";
