+++ editor/editor_help.cpp
@@ -30,6 +30,7 @@
 
 #include "editor_help.h"
 
+#include "configs/modules_enabled.gen.h"
 #include "core/os/input.h"
 #include "core/os/keyboard.h"
 #include "doc_data_compressed.gen.h"
@@ -37,7 +38,6 @@
 #include "editor_node.h"
 #include "editor_scale.h"
 #include "editor_settings.h"
-#include "modules_enabled.gen.h"
 
 #define CONTRIBUTE_URL "https://docs.godotengine.org/en/latest/community/contributing/updating_the_class_reference.html"
