+++ platform/windows/os_windows.cpp
@@ -33,9 +33,9 @@
 
 #include "os_windows.h"
 
+#include "configs/version_generated.gen.h"
 #include "core/io/marshalls.h"
 #include "core/math/geometry.h"
-#include "core/version_generated.gen.h"
 #include "drivers/dummy/rasterizer_dummy.h"
 #include "drivers/gles2/rasterizer_gles2.h"
 #include "drivers/gles3/rasterizer_gles3.h"
