+++ scene/resources/dynamic_font.h
@@ -31,7 +31,7 @@
 #ifndef DYNAMIC_FONT_H
 #define DYNAMIC_FONT_H
 
-#include "modules_enabled.gen.h"
+#include "configs/modules_enabled.gen.h"
 #ifdef MODULE_FREETYPE_ENABLED
 
 #include "core/io/resource_loader.h"
